General refactoring for release management Python scripts
Folders tree reorganization and duplicated code removal
Release notes file ("-current.md") proper cleaning
Removed README.md file update/reset
