+++ README.md
@@ -1,4 +1,4 @@
-Zend 4.0.99
+Zend
 ================
 What is Horizen?
 ----------------
