+++ contrib/devtools/release_management/release_preparation_config.yaml
@@ -1,4 +1,4 @@
-# THIS IS A TEMPLATE FILE, FOR EXAMPLES ON ACTUAL RELEASES CHECK CONTENT OF prepare_release-X.Y.Z.yaml FILES
+# THIS IS A TEMPLATE FILE, FOR EXAMPLES ON ACTUAL RELEASES CHECK CONTENT OF configs/release_preparation_config-X.Y.Z.yaml FILES
 # path to zend project root
 repository_root: "./"
 # new release version (format "major.minor.patch", suffix must not be included if beta or candidate)
@@ -81,12 +81,6 @@ script_steps:
     stop: False
   # update_release_notes step
   update_release_notes:
-    # skip step
-    skip: False
-    # stop before step (user interaction for proceeding)
-    stop: False
-  # update_readme step
-  update_readme:
     # skip step
     skip: False
     # stop before step (user interaction for proceeding)
